@@ -1,14 +1,24 @@
-import React from "react";
-import ReactDOM from "react-dom";
-
-import './index.css';
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import { Provider } from 'react-redux';
+import { store } from './app/store';
 import App from './App';
+import axios from 'axios';
+import './index.css';
 import { ContextProvider } from "./contexts/ContextProvider";
 
-ReactDOM.render(
-  <ContextProvider>
-    <App/>
-  </ContextProvider>
-  ,
-  document.getElementById('root')
+axios.defaults.withCredentials = true;
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <ContextProvider>
+        <App/>
+      </ContextProvider>
+    </Provider>
+  </React.StrictMode>
 );
+
